requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
requests>=2.32.0
beautifulsoup4>=4.13.0
pypdf>=5.0.0
python-docx>=1.1.0
python-multipart>=0.0.12
pillow>=11.0.0
pytesseract>=0.3.13
pytest>=8.3.0
httpx>=0.28.0
